Initial implementation of foldTransversally.
Ideally, this function will subsume all transversal folds,
offering a nice interface to define them.

In order to write foldTransversally in as general a way
as possible, a lone value/left adjoint functor
pseudo-class was defined, in lieu of an official
implementation of the concept somewhere in the
neighbourhood of Data.Functor.Adjunction.