Rating updates within a single race are no longer batched.
Handling races as a collection of one-on-one matches
treats the individual match results as if they were
uncorrelated, which isn't the case, as the match results
of a racer in a race are bound to their laptime. Updating
the ratings immediatly after computing each racing delta
accounts for this correlation to some extent at least.

Note that the list of matches in a race is now transposed
before being fed to the engine core. That, in effect,
gives more weight to matches against racers close to each
other in the scoreboard.