Using the performance model to simulate races and evaluate their strength.
I'm betting on the difficulty of reaching a top-n position against the field
of racers being a better way to measure strength than relying on pairwise
comparisons alone, or just looking at easy-to-calculate victory likelihoods.

There are a few module tidying tweaks here and there in addition to the
simulation implementation. The simulation engine itself still is pretty much
experimental.