Initial experiments with alternative ways of generating combinations.
The eventual goal is generating the tree of combinations and then consuming it
on the fly in a way that reduces the overall number of arithmetical operations
needed to compute the integrals. That is still some way ahead, and there are
no guarantees it will be in fact a big boost.

The combinations code was moved to its own module, Util.Combinations .