Tidying of race ranks is now done during the broader data preparation.
A long overdue change, which was unavoidable now that the remoteness
scores in the engine depend on the actual ranks for the players.

Note that ranks become floating point once data preparation is done. A
Rank' type synonym has been introduced so that their occurrences can be
more easily spotted.