Unify the rating and performance models
The Elo engine, the simulation engine and the top-N probability
algorithm now use the same underlying model. That is a major
simplification, which puts the overall approach on a clearer footing
conceptually, eliminates several ad hoc adjustments and makes it
unecessary to perform interpolation to convert Elo ratings to gamma
rate parameters.

For the moment, the gamma shape parameter of the underlying model is
fixed at 1. On the one hand, that matches exactly the Elo formula on the
ratings side. On the other hand, the resulting performance model is
arguably inferior, in qualitative terms, to the one with shape 3 that
had been formerly used. In any case, the current plan is making the
shape configurable, making it possible to evaluate what is the
ultimately better choice.