The orbitalDistr smart constructor takes ratings
Besides largely doing away with the need for code downstream from
Orbital to deal with the gamma parameters directly, this change makes it
easier to modify how the rating-to-gamma conversion is done.

An analogous change was done to deltaWP, so that the eloAlpha conversion
factor no longer needs to be used outside of Orbital.