Adjust gamma rates according to the shape
The adjustment aims to make equal rating gaps correspond to
approximately equal winning probabilities across gamma shapes. Besides
making it easier to make sense of rating values, the adjustment also
preserves about as much as possible the effect of the Elo modulation
factor.

This commit concludes the main part of the model unification.