Initial prototype of display delta calculation
By "display delta" I mean the difference between rounded down (that is,
as in the published version of the rankings) ratings from one round to
the next.

There are many inconvenient things about how this calculation is
implemented right now. In particular, there is no pleasant way of
applying the activity cut after the processing.